Structure: move json implicits to common
