+++ src/main/scala/com/github/scalalab3/logs/http/AbstractHttpService.scala
@@ -2,7 +2,7 @@ package com.github.scalalab3.logs.http
 
 import akka.actor.ActorRef
 import akka.util.Timeout
-import com.github.scalalab3.logs.json.LogJsonImplicits._
+import com.github.scalalab3.logs.common.json.LogJsonImplicits._
 import com.github.scalalab3.logs.services.{AbstractResponse, BadRequest, LogsResponse, PageLogsResponse}
 import spray.http.HttpHeaders.RawHeader
 import spray.http.{HttpEntity, HttpResponse, StatusCodes}
